third_party/blink/web_tests/external/wpt/css/css-properties-values-api/unit-cycles-expected.txt

This is a testharness.js-based test.
[FAIL] Registered universal variables referencing font-dependent units can be used in font-size
  assert_equals: expected "32px" but got "16px"
Harness: the test ran to completion.
